+++ hoki/csp/tests/test_utils.py
@@ -1,4 +1,6 @@
 """
+Author: Max Briel
+
 Test for the CPS utility subpackages
 """
 import os
@@ -13,57 +15,9 @@ import pytest
 import hoki.csp.utils as utils
 import hoki.load
 from hoki.constants import *
-from hoki.csp.sfh import SFH
 from hoki.utils.exceptions import HokiFormatError, HokiKeyError, HokiTypeError
 
 data_path = pkg_resources.resource_filename('hoki', 'data')
-
-#################################################
-# Test Complex Stellar Populations Parent Class #
-#################################################
-
-
-class TestCSP(object):
-
-    # Test `now` attribute
-    def test_init(self):
-        csp = utils.CSP()
-        assert csp.now == HOKI_NOW, "CSP parent class initialisation failed."
-
-    def test_type_check_history(self):
-        csp = utils.CSP()
-
-        # define pyton callable
-        def x(i): return i
-
-        # define SFH object
-        time_axis = np.linspace(0, 13e9, 1000)
-        sfh = SFH(time_axis, "b", {"constant": 10, "T0": 5e9})
-
-        # Check Types
-        with pytest.raises(HokiTypeError):
-            csp._type_check_histories([10], [0])
-        with pytest.raises(HokiTypeError):
-            csp._type_check_histories([10], 0)
-        with pytest.raises(HokiTypeError):
-            csp._type_check_histories([x, x], [x, 10])
-
-        # Check Format
-        with pytest.raises(HokiFormatError):
-            csp._type_check_histories([x], [x, x])
-        with pytest.raises(HokiFormatError):
-            csp._type_check_histories([x, x], [x])
-        with pytest.raises(HokiFormatError):
-            csp._type_check_histories([x], [])
-
-        # Checking if the correct input does run
-        assert csp._type_check_histories([x], [x]) == ([x], [x])
-        assert csp._type_check_histories([x, x], [x, x]) == ([x, x], [x, x])
-        assert csp._type_check_histories(x, x) == ([x], [x])
-        assert csp._type_check_histories([x], x) == ([x], [x])
-        assert csp._type_check_histories(sfh, x) == ([sfh], [x])
-        assert csp._type_check_histories([sfh], [x]) == ([sfh], [x])
-
 
 #############################
 # Test Calculations per bin #
