+++ hoki/csp/utils.py
@@ -1,4 +1,6 @@
 """
+Author: Max Briel
+
 Utilities to be used in the complex stellar populations
 """
 
@@ -8,104 +10,8 @@ import pandas as pd
 
 import hoki.load
 from hoki.constants import *
-from hoki.csp.sfh import SFH
 from hoki.data_compilers import SpectraCompiler
 from hoki.utils.exceptions import *
-
-############################################
-# Complex Stellar Populations Parent Class #
-############################################
-
-
-class CSP(object):
-    """
-    Complex Stellar Population class
-
-    Notes
-    -----
-    Parent class for `CSPEventRate` and `CSPSpectra`
-
-    Attributes
-    ----------
-    now : `float`
-        The age of the universe.
-    """
-    now = HOKI_NOW
-
-    def __init__(self):
-        pass
-
-    def _type_check_histories(self, sfh, zeh):
-        """
-        Function to make sure inputted stellar formation history functions and
-        metallicity histories are in the correct format.
-
-        Input
-        -----
-        sfh
-            Object to check the type of
-        zeh
-            Object to check the type of
-
-        Returns
-        -------
-        `tuple` ([sfh], [zeh])
-            A tuple containing the sfh callables and zeh callables as arrays.
-        """
-        # Check sfh list
-        if isinstance(sfh, list):
-
-            # check if zeh is also a list
-            if isinstance(zeh, list):
-                # have to be equal lengths
-                if len(sfh) == len(zeh):
-                    # have to be all callables
-                    if (all(callable(val) for val in sfh) and all(callable(val) for val in zeh)):
-                        return (sfh, zeh)
-                    # A non-callable is present
-                    else:
-                        raise HokiTypeError(
-                            "SFH or ZEH contains an object that's not a SFH object or function.")
-                # sfh and zeh are not equal length
-                else:
-                    raise HokiFormatError(
-                        "sfh_functions and Z_functions must have the same length.")
-            # zeh is not a list
-            else:
-                # sfh has to be 1 length and a callable, zeh has to be a callable
-                if len(sfh) == 1:
-                    if (callable(sfh[0]) and callable(zeh)):
-
-                        return (sfh, [zeh])
-                    else:
-                        raise HokiTypeError(
-                            "SFH or ZEH contains an object that's not a SFH object or python callable.")
-                else:
-                    raise HokiFormatError(
-                        "SFH must have length 1, be a python callable, or a SFH object.")
-        # sfh is a callable
-        elif callable(sfh):
-            # zeh is a list
-            if isinstance(zeh, list):
-                # list has to be 1 long, because sfh is callable
-                if len(zeh) == 1:
-                    return ([sfh], zeh)
-                # list it too long
-                else:
-                    raise HokiFormatError(
-                        "ZEH must be either length 1 or a python callable")
-
-            # zeh is also a callable return
-            elif callable(zeh):
-                return ([sfh], [zeh])
-            else:
-                raise HokiTypeError(
-                    "ZEH is not a python callable or a list of callables.")
-        # sfh cannot be identified
-        else:
-            raise HokiTypeError(
-                "SFH type is not a python callable or a SFH object."
-            )
 
 ########################
 # Calculations per bin #
