+++ hoki/tests/test_data_compilers.py
@@ -1,4 +1,6 @@
 """
+Author: Max Briel
+
 Tests for the data_compiler package
 """
